feat: add profile endpoint
