+++ app/Http/Controllers/AuthController.php
@@ -1,15 +1,37 @@
 <?php
 
-namespace App\Http\Controllers\Auth;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use App\Http\Requests\Auth\SignInRequest;
+use App\Http\Requests\Auth\SignUpRequest;
 use App\Models\Player;
 use Illuminate\Support\Facades\Hash;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 
-class SignInController extends Controller
+class AuthController extends Controller
 {
+    /**
+     * Create an account.
+     *
+     * @param  \App\Http\Requests\Auth\SignUpRequest  $request
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function signUp(SignUpRequest $request)
+    {
+        $validated = $request->validated();
+
+        $validated['password'] = bcrypt($validated['password']);
+
+        $player = Player::create($validated);
+
+        $token = $player->createToken('sign-in')->plainTextToken;
+
+        return response()->json([
+            'data' => $player,
+            'token' => $token,
+        ], 201);
+    }
+
     /**
      * Sign in to system.
      *
